koji_promote: remove obsolete koji_promote plugin
Fixes OSBS-3863

The koji_promote plugin is now obsolete, so remove it entirely.
